Fill out classes and enums.  Add DUNSNumber handling to InternalAssociate

+++ BusinessAssociates.Domain/ValueObjects/InternalAssociateId.cs
@@ -4,17 +4,15 @@ namespace BusinessAssociates.Domain.ValueObjects
 {
     public class InternalAssociateId : Value<InternalAssociateId>
     { 
-        public long Value { get; }
+        public int Value { get; }
 
-        public InternalAssociateId(long value)
+        public InternalAssociateId(int value)
         {
             Value = value;
         }
 
-        public static implicit operator long(InternalAssociateId self) => self.Value;
-
-        public static implicit operator InternalAssociateId(long value)
-            => new InternalAssociateId(value);
+        public static implicit operator int(InternalAssociateId self) => self.Value;
+        public static implicit operator InternalAssociateId(int value) => new InternalAssociateId(value);
 
         public override string ToString() => Value.ToString();
     }
